--- Assignment-2/code/my-app/src/App.js
@@ -1,18 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-
-
-
-function App(props) {
- const {employee}=props;
-    
-return(
-  <div className="container">
-  {/* Start your React code here */}
-</div>
-)
-}
-
-
-
-export default App;
