+++ Assignment-2/code/src/index.css
@@ -11,3 +11,7 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+.root{
+  width: 100vw;
+  height: 100vh;
+}
